Pq o MJ atravessou a rua?
Code from the game based on "Freeway" from Activision. Instead of chicken, I put Michael Jackson; the title it's a question that don't have a specific aswer, be free to imagine why he did crossing. When the car hits him, he made the famous "OW" (or "AU" in brazilian portuguese); and when he arrived on the otherside from the road, he make the another famous sound "HEE HEE" (or "IRRI" in brazilian portuguese).